Made `Message` understand `providerStates`.
`Message` previously did not implemented the `providerStates` field as
expected for Pact's V3 specification. This meant the derived
`Deserialize` and `Serialize` implementations were silently wrong,
and that the Pact reference would fail to understand valid files which
used that feature while configured for the V3 specification.

This commit changes the field name on `Message` to correct serialization
and deserialization, updates the `from_json` impl to use
`ProviderState::from_json`, and makes a small style fix to the
initialization of the `Message` struct at the end of the `from_json`
function.

Side note: it's unclear why `from_json` (which converts from a
generic `serde_json::Value` struct) and `Deserialize` (which converts
from the raw read data into the `Message` struct) are both provided.
Quick examination suggests the `Deserialize` impl is vestigial, but it
is left here (and fixed!) to avoid more breakage than is necessary.

Note that because the `providerStates` field name is public and the
`Serialize`/`Deserialize` impls are present, this is a breaking change.