Fix RequestFilterExecutor w/ verify_provider
This fixes a compilation failure in `pact_verifier_cli` due to a
typechecking issue with a `None` variant of an `Option<F> where F:
RequestFilterExecutor` by introducing a `NullRequestFilterExecutor`
and annotating the `None` variant to have the type
`Option<Box<NullRequestFilterExecutor>>`. This allows the crate to
compile as expected, and does not change behavior.

Note that if `NullRequestFilterExecutor` is ever attempted to be called
(unlike the current usage, where it is solely used as a type annotation)
it will panic the current thread.