At least partially correct broken Serialize/Deserialize for Message.
The `Serialize`/`Deserialize` impls for a number of types here aren't
compliant with the spec. This commit attempts to at least partially
correct that for the `Message` type and the types of some of its fields.
This raises the bigger questions of whether this is a desirable
approach (using `Serialize`/`Deserialize`) or it it's better to use
the current preferred process of 1) deserializing into
`serde_json::Value` and then 2) converting with a `from_json`
constructor into the desired type.

Personally I think using `Serialize`/`Deserialize` (possibly needing at
times to do custom trait impls, rather than derives) is better, but I
don't yet have a sense of the level of effort. That would obviously
need to be weighed carefully against the payoff.

If, on the other hand, it's considered better to stick with the current
approach, then the derives ought to be removed, as they're currently
silently wrong, both failing to accept valid inputs and producing
invalid outputs, per the spec.