feat(ffi): add `add_text_comment`
While the addition of `set_comment` allows for the comments to be used
as per the Pact specification, it did not follow the original intent for
comments (see the discussion in pact-foundation/pact-compatibility-suite#8).

The new `add_text_comment` function allows comments to be added
repeatedly and appended into an array as intended in
pact-foundation/pact-specification#45.

Signed-off-by: JP-Ellis <josh@jpellis.me>